# More efficient root path look up (#3302)