**Reasoning**
Most http apps don't use custom http methods. In fact most probably only use GET, POST, PUT and DELETE.
Having an efficient lookup for the default methods, seems the most reasonable solution.

Current impl.
```
[info] Benchmark                              Mode  Cnt      Score       Error  Units
[info] MethodLookupBenchmark.customMethods   thrpt    3  17133.273 ± 13370.029  ops/s
[info] MethodLookupBenchmark.defaultMethods  thrpt    3  18215.269 ± 17379.965  ops/s
```

Map instead of ListMap
```
[info] Benchmark                              Mode  Cnt      Score      Error  Units
[info] MethodLookupBenchmark.customMethods   thrpt    3  22974.932 ± 1760.853  ops/s
[info] MethodLookupBenchmark.defaultMethods  thrpt    3  25215.348 ± 4809.175  ops/s
```

New filed per default HTTP method
```
[info] Benchmark                              Mode  Cnt      Score      Error  Units
[info] MethodLookupBenchmark.customMethods   thrpt    3  31862.950 ± 3407.335  ops/s
[info] MethodLookupBenchmark.defaultMethods  thrpt    3  32393.768 ±  765.915  ops/s

```

fixes #3302
/claim #3302